requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

auraloss
librosa==0.8.1
matplotlib==3.5.2
numpy==1.23.5
python_dateutil==2.8.2
scipy==1.6.1
torch==2.0.0
torchaudio==2.0.0
tqdm==4.62.3
jupyterlab==4.0.2